+++ src/egeo-demo/st-input-demo/st-input-demo.ts
@@ -13,14 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { Component } from '@angular/core';
 
-export { StAlertsDemoModule } from './st-alert-demo/st-alerts-demo.module';
-export {
-   StTwoListSelectionDemoModule
-} from './st-two-list-selection-demo/st-two-list-selection-demo.module';
-export { StTooltipDemoModule } from './st-tooltip-demo/st-tooltip-demo.module';
-export { StSwitchDemoModule } from './st-switch-demo/st-switch-demo.module';
-export {
-   StBreadcrumbsDemoModule
-} from './st-breadcrumbs-demo/st-breadcrumbs-demo.module';
-export { StInputDemoModule } from './st-input-demo/st-input-demo.module';
+@Component({
+   selector: 'st-input-demo',
+   templateUrl: 'st-input-demo.html'
+})
+export class StInputDemoComponent {
+}